separate out percentage chart
