+++ locales-pending/premium.ftl
@@ -305,7 +305,7 @@ data-broker-profiles-exposure-reduction = Exposure reduction: { $exposure_reduct
 # the label inside <label> will be shown in a smaller font.
 # Variables:
 #   $nr (number) - % of exposures reduced for the user
-exposure-reduction-chart-heading = <nr>{ $nr }</nr><percent>%</percent> <label1>exposures will</label1><label2>be reduced</label2>
+exposure-reduction-chart-heading = <nr>{ $nr }</nr><percent>%</percent> <label_line_1>exposures will</label_line_1><label_line_2>be reduced</label_line_2>
 
 welcome-to-premium-data-broker-profiles-title-part-one = Welcome to Premium.
 welcome-to-premium-data-broker-profiles-title-part-two = We’ll remove those profiles ASAP.
